+++ src/account/builder.rs
@@ -106,7 +106,10 @@ impl AccountBuilder {
             let account = account_handle.read().await;
             let existing_coin_type = account.coin_type;
             if existing_coin_type != self.coin_type {
-                return Err(Error::InvalidCoinType(self.coin_type, existing_coin_type));
+                return Err(Error::InvalidCoinType {
+                    new_coin_type: self.coin_type,
+                    existing_coin_type,
+                });
             }
             if account.alias().to_lowercase() == account_alias.to_lowercase() {
                 return Err(Error::AccountAliasAlreadyExists(account_alias));
@@ -185,6 +188,7 @@ impl AccountBuilder {
             transactions: HashMap::new(),
             pending_transactions: HashSet::new(),
             incoming_transactions: HashMap::new(),
+            native_token_foundries: HashMap::new(),
         };
 
         let account_handle = AccountHandle::new(